+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## v.NEXT
 
+**Bug fixes**
+
+* Define `ColonyClient.addDomain` as a `MultisigSender` (`@colony/colony-js-client`)
+* Define `ColonyClient.cancelTask` as a `MultisigSender` (`@colony/colony-js-client`)
 
 ## v1.8.0